+++ terraform/gce.sample.tf
@@ -1,14 +1,15 @@
 variable "control_count" { default = 3 }
+variable "control_type" { default = "n1-standard-1" }
 variable "datacenter" {default = "gce"}
 variable "edge_count" { default = 1}
+variable "edge_type" { default = "n1-standard-1" }
 variable "image" {default = "centos-7-v20160119"}
+variable "kube_worker_count" {default = 0}
 variable "long_name" {default = "mantl"}
 variable "short_name" {default = "mi"}
 variable "ssh_key" {default = "~/.ssh/id_rsa.pub"}
 variable "ssh_user" {default = "centos"}
 variable "worker_count" {default = 4}
-variable "control_type" { default = "n1-standard-1" }
-variable "edge_type" { default = "n1-standard-1" }
 variable "worker_type" { default = "n1-standard-2" }
 variable "zones" {
   default = "us-central1-a,us-central1-b"
@@ -84,6 +85,21 @@ module "worker-nodes" {
   zones = "${var.zones}"
 }
 
+module "kube-worker-nodes" {
+  source = "./terraform/gce/instance"
+  count = "${var.kube_worker_count}"
+  datacenter = "${var.datacenter}"
+  image = "${var.image}"
+  machine_type = "n1-highmem-2"
+  network_name = "${module.gce-network.network_name}"
+  #network_name = "${terraform_remote_state.gce-network.output.network_name}"
+  role = "kubeworker"
+  short_name = "${var.short_name}"
+  ssh_user = "${var.ssh_user}"
+  ssh_key = "${var.ssh_key}"
+  zones = "${var.zones}"
+}
+
 module "network-lb" {
   source = "./terraform/gce/lb"
   instances = "${module.edge-nodes.instances}"
@@ -103,4 +119,6 @@ module "cloud-dns" {
   subdomain = ".service"
   worker_count = "${var.worker_count}"
   worker_ips = "${module.worker-nodes.gce_ips}"
+  kubeworker_count = "${var.kube_worker_count}"
+  kubeworker_ips = "${module.kube-worker-nodes.gce_ips}"
 }